// apps/humayro/lib/api/model/organizationAdvertisementResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Humayro API
 * Humayro e-commerce API
 * OpenAPI spec version: 1.0
 */

export interface OrganizationAdvertisementResponse {
  id?: number;
  organizationId?: number;
  organizationName?: string;
  videoAttachmentId?: number;
  videoUrl?: string;
  title?: string;
  redirectUrl?: string;
  active?: boolean;
  sortOrder?: number;
  startsAt?: string;
  endsAt?: string;
  createdAt?: string;
  updatedAt?: string;
}